Refactor order estimation to return float, update admin dashboard access control, and enhance cart functionality with improved error handling and user feedback. Added integration tests for authentication and cart operations, and removed outdated test files. Updated email utility tests for better coverage.

+++ models/order.py
@@ -23,7 +23,7 @@ class Order(db.Model):
         for item in self.items:
             total += item.quantity * item.product.price
         
-        return round(total,2)
+        return float(round(total,2))
     
     def complete(self):
         products_to_update = []
